Destini: TODO's Completed:
Step 9 - Create a new storyBrain object from the StoryBrain class.

Step 10 - use the storyBrain to get the first story title and display it in this Text Widget.

Step 11 - Create a method called getChoice1() that returns the text for the first choice1 from _storyData.

Step 12 - Create a method called getChoice2() that returns the text for the first choice2 from _storyData.